api/pyproject.toml
[tool.poetry]
name = "construction"
version = "0.1.0"
description = ""
authors = ["gcca <git@gcca.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
markdown = "^3.6"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
coverage = "^7.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
